Restructure according to Tommer's feedback
