+++ frontend/src/components/Curator/CuratorInput.jsx
