+++ frontend/src/components/Curator/CuratorSettingsDrawer.jsx
