+++ frontend/src/components/Curator/views/CuratedView.jsx
