+++ frontend/src/components/Curator/views/ExportedView.jsx
